grpc-trial/requirements.txt
black==21.9b0
click==8.0.1
grpcio==1.41.0
grpcio-tools==1.41.0
mypy-extensions==0.4.3
mypy-protobuf==2.10
numpy==1.21.2
pathspec==0.9.0
platformdirs==2.4.0
protobuf==3.18.1
regex==2021.9.30
scipy==1.7.1
six==1.16.0
tomli==1.2.1
types-futures==3.3.0
types-protobuf==3.17.4
typing-extensions==3.10.0.2
